+++ pybamm/input/parameters/lithium_ion/electrolytes/lipf6_EC_DMC_1_1_Landesfeind2019/electrolyte_conductivity_EC_DMC_1_1_Landesfeind2019.py
@@ -1,4 +1,4 @@
-from electrolyte_base_Landesfeind2019 import (
+from .electrolyte_base_Landesfeind2019 import (
     electrolyte_conductivity_base_Landesfeind2019,
 )
 import numpy as np